sigops.scala
